Testing remote codeBase setup, using an overlay.
File created locally as a Framer overlay with the intent to import it into the overlay script in Framer and then update it on change. This way the entire Framer coding process can be moved externally and worked on from a local repo. Plus those fuckers can't delete your work for spite.